# swaggerui/swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/convert": {
      "post": {
        "description": "Converts a supplied K8S Secret manifest (JSON) to an Encrypted Secret",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "convert"
        ],
        "operationId": "Convert",
        "parameters": [
          {
            "description": "Existing K8S Secret in JSON format.",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/kubeSecret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversion successful"
          },
          "400": {
            "description": "Invalid request body"
          },
          "500": {
            "description": "Conversion Error"
          }
        }
      }
    },
    "/create": {
      "post": {
        "description": "Creates a new encrypted secret object from raw data",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "create"
        ],
        "operationId": "Create",
        "parameters": [
          {
            "description": "Details of secret to be created",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/secretRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Creation successful"
          },
          "400": {
            "description": "Invalid request body"
          },
          "500": {
            "description": "Creation error"
          }
        }
      }
    }
  },
  "definitions": {
    "kubeSecret": {
      "description": "kubeSecret",
      "type": "object",
      "required": [
        "kind",
        "apiVersion",
        "data"
      ],
      "properties": {
        "apiVersion": {
          "type": "string",
          "x-go-name": "APIVersion",
          "example": "v1"
        },
        "data": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "uint8"
            }
          },
          "x-go-name": "Data",
          "example": {
            "admin": "YWRtaW4=",
            "password": "cGFzc3dvcmQ="
          }
        },
        "kind": {
          "type": "string",
          "x-go-name": "Kind",
          "example": "Secret"
        },
        "metadata": {
          "$ref": "#/definitions/secretMetadata"
        }
      },
      "x-go-name": "KubeSecret",
      "x-go-package": "gitlab.com/sgryczan/seel/api"
    },
    "secretMetadata": {
      "description": "Secret Metadata",
      "type": "object",
      "required": [
        "name",
        "namespace"
      ],
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "myAwesomeSecret"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace",
          "example": "destinationNamespace"
        }
      },
      "x-go-name": "SecretMetadata",
      "x-go-package": "gitlab.com/sgryczan/seel/api"
    },
    "secretRequest": {
      "description": "Secret Request",
      "type": "object",
      "required": [
        "name",
        "namespace",
        "data"
      ],
      "properties": {
        "data": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Data",
          "example": {
            ".dockerconfigjson": "eyJhdXRocyI6eyJodHRwczovL2luZGV4LmRvY2tlci5...",
            "admin": "my-serviceaccount",
            "certificate": "InBhc3N3b3JkIjoiSWJjdDk0MjAiLCJlbWFpbCI6InNlYmFz...",
            "password": "start123"
          }
        },
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "my-secret"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace",
          "example": "app-namespace"
        }
      },
      "x-go-name": "SecretRequest",
      "x-go-package": "gitlab.com/sgryczan/seel/api"
    }
  }
}